backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.39.0
pymupdf==1.24.0
python-dotenv==1.0.1
pydantic==2.9.0
requests==2.31.0